backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
xgboost>=2.0.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
shap>=0.44.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
pymongo>=4.6.0
motor>=3.5.0
python-multipart>=0.0.9
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
